Move redefined C structures to tun.h
Reuse fd in legacy from NETunnelInterface.
