+++ Sources/PartoutCore_C/include/portable/tun.h
@@ -13,6 +13,23 @@
 #include "portable/common.h"
 #include "portable/socket.h"
 
+/* Redefine these manually because the <sys/kern_control.h>
+ * header is not exposed to iOS/tvOS */
+#if PARTOUT_APPLE && !PARTOUT_MACOS
+struct ctl_info {
+    u_int32_t   ctl_id;
+    char        ctl_name[96];
+};
+struct sockaddr_ctl {
+    u_char      sc_len;
+    u_char      sc_family;
+    u_int16_t   ss_sysaddr;
+    u_int32_t   sc_id;
+    u_int32_t   sc_unit;
+    u_int32_t   sc_reserved[5];
+};
+#endif
+
 #pragma clang assume_nonnull begin
 
 /* Opaque tun device. */
